feat: migrate project to typescript
