+++ src/app.ts
@@ -1,11 +1,11 @@
-const Koa = require('koa');
-const koaBody = require('koa-body');
+import Koa from 'koa';
+import koaBody from 'koa-body';
 // const swaggerUI = require('swagger-ui-express');
 // const path = require('path');
 // const YAML = require('yamljs');
-const userRouter = require('./resources/users/user.router');
-const boardRouter = require('./resources/boards/board.router');
-const taskRouter = require('./resources/tasks/task.router');
+import userRouter from './resources/users/user.router';
+import boardRouter from './resources/boards/board.router';
+import taskRouter from './resources/tasks/task.router';
 
 const app = new Koa();
 // const swaggerDocument = YAML.load(path.join(__dirname, '../doc/api.yaml'));
@@ -18,4 +18,4 @@ app.use(userRouter.routes());
 app.use(boardRouter.routes());
 app.use(taskRouter.routes());
 
-module.exports = app;
+export default app;
