--- src/resources/users/user.model.js
@@ -1,22 +1,0 @@
-const uuidv4 = require('uuid').v4;
-
-class User {
-  constructor({
-    id = uuidv4(),
-    name = 'USER',
-    login = 'user',
-    password = 'P@55w0rd'
-  } = {}) {
-    this.id = id;
-    this.name = name;
-    this.login = login;
-    this.password = password;
-  }
-
-  toResponse() {
-    const { id, name, login } = this;
-    return { id, name, login };
-  }
-}
-
-module.exports = User;
