--- src/server.js
@@ -1,12 +1,0 @@
-const { PORT } = require('./common/config');
-const app = require('./app');
-
-const _PORT = PORT || 4000;
-
-app.listen(_PORT,'localhost', (e) => {
-if (e){
-    console.log(e)
-} else {
-    console.log(`Server is listening on port ${_PORT}`)
-}
-});
